requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
astrapy
openai
tiktoken
pyasyncore